Getting started with implementing DTOs

+++ bitquiz/src/main/java/com/bitconex/bitquiz/dto/QuizQuestionsDTO.java
@@ -12,5 +12,5 @@ public class QuizQuestionsDTO {
     private String text;
     private int points;
     private boolean helpAllowed;
-    private long quizId;
+    private int quizId;
 }
